+++ assets/javascripts/kitten/components/grid/container/stories.js
@@ -1,0 +1,5 @@
+import React from 'react'
+import { storiesOf } from '@storybook/react'
+import { Container } from './index'
+
+storiesOf('Grid/Container', module).add('default', <Container />)
